protecting contacts routes from unauthorized user

+++ controllers/contactController.js
@@ -54,6 +54,11 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("Contact Not found");
   }
 
+  if(contact.user_id.toString() !== req.user.id){
+    res.status(403)
+    throw new Error("user don't have permission to update others contacts");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -73,6 +78,12 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact Not found");
   }
+
+  if(contact.user_id.toString() !== req.user.id){
+    res.status(403)
+    throw new Error("user don't have permission to update others contacts");
+  }
+
 
   await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json(contact);
